requirements.txt
# Web framework + scheduler
fastapi==0.115.5
uvicorn[standard]==0.32.1
apscheduler==3.10.4
jinja2==3.1.4
itsdangerous==2.2.0
python-multipart==0.0.20

# Config + logging
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
structlog==24.4.0

# Microsoft Graph auth + HTTP
msal==1.31.1
httpx==0.28.1
tenacity==9.0.0

# Document processing (Unstructured self-hosted)
# Pin to a known-good version with all extras
unstructured[all-docs]==0.16.11
# Explicit deps that 'all-docs' brings in but worth pinning:
pdfminer.six==20231228
pi-heif==0.21.0
unstructured-inference==0.8.1
pikepdf==9.4.2

# Embeddings + vector DB
openai==1.57.4
pinecone-client==5.0.1

# Crypto for at-rest secret encryption
cryptography==44.0.0

# Misc
tqdm==4.67.1
